refactor to share common mpu code between i2c and SPI
